+++ src/main/kotlin/ch/daplab/nosql/cassandra/doodle/services/impl/cassandra/CassandraPollServiceImpl.kt
@@ -1,20 +1,22 @@
 package ch.daplab.nosql.cassandra.doodle.services.impl.cassandra
 
+import ch.daplab.nosql.cassandra.doodle.services.impl.CassandraHelpers
+import ch.daplab.nosql.cassandra.doodle.services.impl.CassandraHelpers.toUUID
 import ch.daplab.nosql.cassandra.doodle.domains.Poll
 import ch.daplab.nosql.cassandra.doodle.domains.Subscriber
 import ch.daplab.nosql.cassandra.doodle.services.PollService
 import com.datastax.driver.core.Row
 import com.datastax.driver.core.Session
 import com.datastax.driver.core.UDTValue
-import com.datastax.driver.core.UserType
 import java.util.*
 
+
 class CassandraPollServiceImpl(val session: Session, autoCreateSchema: Boolean = true) : PollService {
-    val subscriberUDT: UserType
+
+    val subscriberUDT = session.cluster.metadata.getKeyspace("doodle").getUserType("subscriber")
 
     init {
-        if (autoCreateSchema) createSchema()
-        subscriberUDT = session.cluster.metadata.getKeyspace("doodle").getUserType("subscriber")
+        if (autoCreateSchema) CassandraHelpers.createSchema(session)
     }
 
     override fun getPollById(pollId: String): Poll? {
@@ -53,35 +55,7 @@ class CassandraPollServiceImpl(val session: Session, autoCreateSchema: Boolean =
     override val allPolls: List<Poll>
         get() = session.execute("SELECT * FROM doodle.polls").map(::rowToPoll)
 
-    fun createSchema() {
-        with (session) {
-            // The keyspace should probably be created elsewhere...
-            execute("""
-                CREATE KEYSPACE IF NOT EXISTS doodle
-                    WITH replication = {'class': 'org.apache.cassandra.locator.SimpleStrategy',
-                        'replication_factor': 3 }
-            """)
-            execute("""
-                CREATE TYPE IF NOT EXISTS doodle.subscriber (
-                    label text,
-                    choices list<text>
-                )
-            """)
-            execute("""
-                CREATE TABLE IF NOT EXISTS doodle.polls (
-                    id UUID PRIMARY KEY,
-                    label text,
-                    choices list<text>,
-                    email text,
-                    maxChoices int,
-                    subscribers list<FROZEN <subscriber>>
-                )
-            """)
-        }
-    }
 }
-
-private fun String.toUUID() = UUID.fromString(this)
 
 internal fun rowToPoll(row: Row): Poll {
     return Poll().apply {
