+++ src/main/kotlin/ch/daplab/nosql/cassandra/doodle/services/impl/cassandraOM/CassandraOMPollServiceImpl.kt
@@ -1,9 +1,10 @@
 package ch.daplab.nosql.cassandra.doodle.services.impl.cassandraOM
 
+import ch.daplab.nosql.cassandra.doodle.services.impl.CassandraHelpers
+import ch.daplab.nosql.cassandra.doodle.services.impl.CassandraHelpers.toUUID
 import ch.daplab.nosql.cassandra.doodle.domains.Poll
 import ch.daplab.nosql.cassandra.doodle.domains.Subscriber
 import ch.daplab.nosql.cassandra.doodle.services.PollService
-import ch.daplab.nosql.cassandra.doodle.services.impl.cassandra.CassandraPollServiceImpl
 import com.datastax.driver.core.Session
 import com.datastax.driver.mapping.Mapper
 import com.datastax.driver.mapping.MappingManager
@@ -32,7 +33,7 @@ class CassandraOMPollServiceImpl(val session: Session, autoCreateSchema: Boolean
     val pollAccessor: PollAccessor
 
     init {
-        if (autoCreateSchema) createSchema()
+        if (autoCreateSchema) CassandraHelpers.createSchema(session)
         manager = MappingManager(session)
         pollMapper = manager.mapper(PollModel::class.java)
         pollAccessor = manager.createAccessor(PollAccessor::class.java)
@@ -71,12 +72,4 @@ class CassandraOMPollServiceImpl(val session: Session, autoCreateSchema: Boolean
     override val allPolls: List<Poll>
         get() = pollAccessor.getAll().map { it.toPoll() }
 
-
-    fun createSchema() {
-        // TODO: Fix this
-        val impl = CassandraPollServiceImpl(session, autoCreateSchema = false)
-        impl.createSchema()
-    }
 }
-
-private fun String.toUUID() = UUID.fromString(this)
